requirements_lambda.txt
playwright
pillow
openai
python-dotenv
boto3
